# refactor: reuse the lua vm and pr-compile the lua code